+++ aitoolman/model.py
@@ -372,7 +372,7 @@ class LLMModuleRequest(typing.NamedTuple):
     """应用层模板请求参数（模块配置）"""
     module_name: str
     template_params: Dict[str, Any]
-    # 模型名称或模型标签（多个取并集）
+    # 模型名称或模型标签（多个取交集）
     model_name: Union[str, List[str], None] = None
     model_rank: int = 0
     context_messages: List[Message] = []
@@ -497,6 +497,12 @@ class LLMModuleRequestState(typing.NamedTuple):
     module_request: Optional[LLMModuleRequest]
     direct_request: LLMDirectRequest
     provider_request: LLMProviderRequest
+
+    def has_response(self) -> bool:
+        return self.provider_request.response.done()
+
+    async def wait_response(self) -> LLMProviderResponse:
+        return await self.provider_request.response
 
 
 __all__ = [